# Variable statement optimization