--- a/Jint/Runtime/Interpreter/Expressions/JintMemberExpression.cs
+++ b/Jint/Runtime/Interpreter/Expressions/JintMemberExpression.cs
@@ -1,4 +1,5 @@
 using Esprima.Ast;
+using Jint.Native;
 using Jint.Runtime.References;
 
 namespace Jint.Runtime.Interpreter.Expressions
@@ -9,12 +10,15 @@ namespace Jint.Runtime.Interpreter.Expressions
     internal sealed class JintMemberExpression : JintExpression
     {
         private readonly JintExpression _objectExpression;
+        private readonly JintIdentifierExpression _objectIdentifierExpression;
+
         private readonly JintExpression _propertyExpression;
         private readonly string _determinedPropertyNameString;
 
         public JintMemberExpression(Engine engine, MemberExpression expression) : base(engine, expression)
         {
             _objectExpression = Build(engine, expression.Object);
+            _objectIdentifierExpression = _objectExpression as JintIdentifierExpression;
             if (!expression.Computed)
             {
                 _determinedPropertyNameString = ((Identifier) expression.Property).Name;
@@ -28,8 +32,36 @@ namespace Jint.Runtime.Interpreter.Expressions
 
         protected override object EvaluateInternal()
         {
-            var baseReference = _objectExpression.Evaluate();
-            var baseValue = _engine.GetValue(baseReference, false);
+            string baseReferenceName = null;
+            JsValue baseValue = null;
+            var isStrictModeCode = StrictModeScope.IsStrictModeCode;
+
+            if (_objectIdentifierExpression != null)
+            {
+                baseReferenceName = _objectIdentifierExpression._expressionName;
+                var strict = isStrictModeCode;
+                TryGetIdentifierEnvironmentWithBindingValue(
+                    strict,
+                    _objectIdentifierExpression._expressionName,
+                    out _,
+                    out baseValue);
+            }
+
+            if (baseValue is null)
+            {
+                // fast check failed
+                var baseReference = _objectExpression.Evaluate();
+                if (baseReference is Reference reference)
+                {
+                    baseReferenceName = reference._name;
+                    baseValue = _engine.GetValue(reference, false);
+                    _engine._referencePool.Return(reference);
+                }
+                else
+                {
+                    baseValue = _engine.GetValue(baseReference, false);
+                }
+            }
 
             string propertyNameString = _determinedPropertyNameString;
             if (propertyNameString == null)
@@ -39,14 +71,9 @@ namespace Jint.Runtime.Interpreter.Expressions
                 propertyNameString = TypeConverter.ToPropertyKey(propertyNameValue);
             }
 
-            TypeConverter.CheckObjectCoercible(_engine, baseValue, (MemberExpression) _expression, baseReference);
+            TypeConverter.CheckObjectCoercible(_engine, baseValue, (MemberExpression) _expression, baseReferenceName);
 
-            if (baseReference is Reference r)
-            {
-                _engine._referencePool.Return(r);
-            }
-
-            return _engine._referencePool.Rent(baseValue, propertyNameString, StrictModeScope.IsStrictModeCode);
+            return _engine._referencePool.Rent(baseValue, propertyNameString, isStrictModeCode);
         }
     }
 }