@sebastienros 2018-12-09
Looks good.

My only concern is to know how far it is from the specification, more importantly the es6 one. As it might be harder to integrate the new logic if there are shortcuts done to the spec. Ideally you were using the new spec while working on that. At least all the unit tests pass which means it can't be far from the original spec.