@lahma 2018-12-09
I think we can always backtrack if needed. So far what I've studied new functionalities this shouldn't be much of an impact.